+++ packages/sdk/src/primitives/deposit.ts
@@ -2,7 +2,6 @@ import { StealthAddress } from "../crypto";
 import { EncodedAsset } from "./asset";
 
 export interface DepositRequest {
-  chainId: bigint;
   spender: string;
   encodedAsset: EncodedAsset;
   value: bigint;